+++ lib/oxc/lint/native.ex
@@ -22,13 +22,6 @@ defmodule OXC.Lint.Native do
     ),
     version: version
 
-  @spec lint(
-          iodata(),
-          String.t(),
-          [String.t()],
-          [{String.t(), String.t()}],
-          [{String.t(), String.t()}],
-          boolean()
-        ) :: {:ok, [map()]} | {:error, [String.t()]}
+  @spec lint(iodata(), String.t(), map()) :: {:ok, [map()]} | {:error, [String.t()]}
   use OXC.Lint.Native.GeneratedStubs
 end